ContributorLists component
