+++ src/components/ProjectContributors.js
@@ -11,6 +11,7 @@ import {
 } from 'lodash'
 
 import AllocationAddForm from './AllocationAddForm'
+import ContributorList from './ContributorList'
 import LoadingProgress from './LoadingProgress'
 import ContributorTile from './ContributorTile'
 import ContributorsEmptyState from './ContributorsEmptyState'
@@ -30,13 +31,6 @@ const ProjectContributors = (props) => {
 
     const [contributors, setContributors] = useState([])
     const [githubContributors, setGithubContributors] = useState([])
-    const [openAddAllocationDialog, setOpenAddAllocationDialog] = useState(false)
-    const [contributorClicked, setContributorClicked] = useState(null)
-    const [searchFilter, setSearchFilter] = useState('')
-
-    const handleAddAllocationClose = (value) => {
-        setOpenAddAllocationDialog(false)
-    }
 
     const {
         data: dataProjectContributors,
@@ -132,179 +126,15 @@ const ProjectContributors = (props) => {
         setContributors(contributors.concat(...dataContributors.getContributors))
     }
 
-    const stringMatchName = (string) => {
-        return string.name.toLowerCase().includes(searchFilter.toLowerCase())
-    }
-
-    const stringMatchGithub = (string) => {
-        return string.github_handle.toLowerCase().includes('https://github.com/'.concat(searchFilter).toLowerCase())
-    }
-
-    const getActiveContributors = () => {
-        if (searchFilter == '') {
-            return activeContributorsAllocated
-        } else {
-            return activeContributorsAllocated.filter(a => {
-                return stringMatchName(a) || stringMatchGithub(a)
-            })
-        }
-    }
-    const activeContributors = getActiveContributors()
-
-    const getUpcomingContributors = () => {
-        if (searchFilter == '') {
-            return upcomingContributorsAllocatedOnly
-        } else {
-            return upcomingContributorsAllocatedOnly.filter(u => {
-                return stringMatchName(u) || stringMatchGithub(u)
-            })
-        }
-    }
-    const upcomingContributors = getUpcomingContributors()
-
-    const getContributorsToAdd = () => {
-        const addContributors = differenceBy(contributors, [...activeContributorsAllocated, ...upcomingContributorsAllocatedOnly], 'id')
-        if (searchFilter == '') {
-            return addContributors
-        } else {
-            return addContributors.filter(a => {
-                return stringMatchName(a) || stringMatchGithub(a)
-            })
-        }
-    }
-    const contributorsToAdd = getContributorsToAdd()
-
-    const addAllocation = (props) => {
-        setOpenAddAllocationDialog(true)
-        setContributorClicked(props.contributor)
-    }
-
-    const renderContributors = (props) => {
-        const {
-            active,
-            contributors,
-            project
-        } = props
-        return contributors.map(c => {
-            return (
-                <Grid item xs={12} md={6}>
-                    <ContributorTile
-                        active={active}
-                        contributor={c}
-                        onAddButton={addAllocation}
-                        project={project}
-                    />
-                </Grid>
-            )
-        })
-    }
-
     return (
-        <Grid container className='ProjectContributors'>
-            <h1>
-                {`${project.name} Contributors`}
-            </h1>
-            <Grid xs={12}/>
-            <Grid item xs={12} sm={5}>
-                <Box
-                    bgcolor='primary.black'
-                    color='primary.light'
-                    borderRadius='borderRadius'
-                    px={5}
-                    py={1}
-                    mb={[2, 5]}
-                >
-                    {
-                        `${activeContributorsAllocated.length} active ${activeContributorsAllocated.length == 1
-                            ? 'contributor'
-                            : 'contributors'
-                        }`
-                    }
-                </Box>
-            </Grid>
-            <Grid xs={12}/>
-            <Grid item xs={12} sm={5}>
-                <TextField
-                    placeholder='Search contributors...' 
-                    onChange={(event) => { setSearchFilter(event.target.value) }}
-                    fullWidth
-                    color='primary'
-                    variant='outlined'
-                    type='search'
-                />
-            </Grid>
-            <Grid item xs={12}>
-                <Box my={[2, 5]}>
-                    <Typography align='left' variant='h5'>
-                        {`Active contributors`}
-                    </Typography>
-                    <Grid container>
-                        {
-                            activeContributors.length != 0
-                                ? renderContributors({
-                                    active: true,
-                                    contributors: activeContributors,
-                                    project: project
-                                })
-                                : <ContributorsEmptyState active/>
-
-                        }
-                    </Grid>
-                </Box>
-                <Box my={[2, 5]}>
-                    <Typography align='left' variant='h5'>
-                        {`Upcoming contributors`}
-                    </Typography>
-                    <Grid container>
-                        {
-                            upcomingContributors.length != 0
-                                ? renderContributors({
-                                    active: true,
-                                    contributors: upcomingContributors,
-                                    project: project
-                                })
-                                : <ContributorsEmptyState active/>
-
-                        }
-                    </Grid>
-                </Box>
-                <hr/>
-            </Grid>
-            <h1>
-                {`Add new contributors to the project`}
-            </h1>
-            {errorGithubContributors
-                ? isNotAContributor()
-                : null
-            }
-            <Grid item xs={12}>
-                <Box>
-                    <Grid container>
-                        {
-                            contributorsToAdd.length != 0
-                                ? renderContributors({
-                                    active: false,
-                                    contributors: contributorsToAdd,
-                                    project: project
-                                })
-                                : <ContributorsEmptyState/>
-                        }
-                    </Grid>
-                </Box>
-                <Box my={[2, 5]} py={5}/>
-            </Grid>
-            <Grid item xs={12}>
-                {
-                    contributorClicked &&
-                    <AllocationAddForm
-                        project={project}
-                        open={openAddAllocationDialog}
-                        onClose={handleAddAllocationClose}
-                        contributor={contributorClicked}
-                    />
-                }
-            </Grid>
-        </Grid>
+        <ContributorList 
+            contributors={contributors}
+            project={project}
+            activeContributorsAllocated={activeContributorsAllocated}
+            upcomingContributorsAllocatedOnly={upcomingContributorsAllocatedOnly}
+            errorGithubContributors={errorGithubContributors}
+            isNotAContributor={isNotAContributor}
+        />
     )
 }
